cfp_reviewer_checker/requirements.txt
python>=3.10
oumi>=0.1.0
aiohttp>=3.9.0
pydantic>=2.0.0
typing-extensions>=4.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.0

